feat : token 인증 구현 완료(게시물 연동)

+++ day4-project/src/main/java/com/example/day4_project/controller/UserController.java
@@ -2,13 +2,18 @@ package com.example.day4_project.controller;
 
 import com.example.day4_project.dto.UserRequestDto;
 import com.example.day4_project.dto.UserResponseDto;
+import com.example.day4_project.global.exception.AppException;
+import com.example.day4_project.global.exception.ErrorCode;
 import com.example.day4_project.global.response.DataResponse;
+import com.example.day4_project.security.user.UserDetailsImpl;
 import com.example.day4_project.service.UserService;
 import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.security.SecurityRequirement;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -36,16 +41,37 @@ public class UserController {
     public ResponseEntity<DataResponse<List<UserResponseDto>>> getAllUsers() {
         return ResponseEntity.ok(DataResponse.of(userService.getAllUsers()));
     }
+
+    @SecurityRequirement(name="bearerAuth")
     @GetMapping("/{id}")
-    public ResponseEntity<DataResponse<UserResponseDto>> getUserById(@PathVariable Long id) {
-        UserResponseDto byId = userService.findById(id);
+    public ResponseEntity<DataResponse<UserResponseDto>> getUserById(@PathVariable Long id, @AuthenticationPrincipal UserDetailsImpl userDetails) {
+        if(!userDetails.getUser().getId().equals(id)) {
+            throw new AppException(ErrorCode.UNAUTHORIZED_REQUEST); // 본인 확인
+        }
+        UserResponseDto byId=userService.findById(id);
         return ResponseEntity.ok(DataResponse.of(byId));
-//        return ResponseEntity.ok(byId);
     }
+    @SecurityRequirement(name = "bearerAuth")
     @DeleteMapping("/{id}")
-    public ResponseEntity<DataResponse<Void>> deleteUserById(@PathVariable Long id) {
+    public ResponseEntity<DataResponse<Void>> deleteUserById(@PathVariable Long id, @AuthenticationPrincipal UserDetailsImpl userDetails){
+        if(!userDetails.getUser().getId().equals(id)) {
+            throw new AppException(ErrorCode.UNAUTHORIZED_REQUEST);
+        }
         userService.delete(id);
         return ResponseEntity.ok(DataResponse.empty());
     }
+
+
+    //security 하기 전
+//    public ResponseEntity<DataResponse<UserResponseDto>> getUserById(@PathVariable Long id) {
+//        UserResponseDto byId = userService.findById(id);
+//        return ResponseEntity.ok(DataResponse.of(byId));
+////        return ResponseEntity.ok(byId);
+//    }
+//    @DeleteMapping("/{id}")
+//    public ResponseEntity<DataResponse<Void>> deleteUserById(@PathVariable Long id) {
+//        userService.delete(id);
+//        return ResponseEntity.ok(DataResponse.empty());
+//    }
 
 }
